PATH PLANNING FOR AUTONOMOUS MOBILE ROBOT.
The goal of mobile robot path planning is to find a path from the current position to the target position. Path planning is one of the most crucial research problems in robotics from the perspective of the control engineer. Many problems in various fields are solved by proposing path planning. It has been applied in guiding the robot to reach a particular objective from very simple trajectory planning to the selection of a suitable sequence of action. By proposing a proper algorithm, path planning can be widely applied in partially and unknown structured environments.mobile robots can find an optimal or near-optimal path from the starting state to the target state that avoids obstacles based on one or some performance indicators (such as the lowest working cost, the shortest walking route, the shortest walking time, etc.) in the motion space. Generally speaking, the path planning selects the shortest distance of the path: that is, the shortest length of the path from the starting point to the target point as a performance indicator. Path planning requires building an environmental map. Environmental map construction refers to the establishment of an accurate spatial location description of various objects in the environment in which the robot is located, including obstacles, road signs, and so on: that is, the establishment of a spatial model or map. The purpose of constructing the environmental map is to help the mobile robot plan an optimal path from the starting point to the target point in the established environment model with obstacles.